+++ tests/views/column-titles-test.blade.php
@@ -1,0 +1,1 @@
+<tr id="column-titles-test"></tr>