+++ tests/views/create-action-test.blade.php
@@ -1,0 +1,3 @@
+<form id="create-action-test"></form>
+
+