+++ tests/views/edit-test.blade.php
@@ -1,3 +1,0 @@
-<form id="edit-test"></form>
-
-