+++ tests/views/navigation-status-test.blade.php
@@ -1,0 +1,1 @@
+<div id="navigation-status-test"></div>