+++ tests/views/pagination-test.blade.php
@@ -1,0 +1,1 @@
+<nav id="pagination-test"></nav>