+++ tests/views/rows-searching-test.blade.php
@@ -1,0 +1,1 @@
+<form id="rows-searching-test"></form>