+++ tests/views/show-test.blade.php
@@ -1,3 +1,0 @@
-<form id="show-test"></form>
-
-